+++ common-client/src/main/java/de/maxhenkel/voicechat/gui/audiodevice/SelectSpeakerScreen.java
@@ -1,20 +1,15 @@
 package de.maxhenkel.voicechat.gui.audiodevice;
 
-import de.maxhenkel.configbuilder.entry.ConfigEntry;
-import de.maxhenkel.voicechat.Voicechat;
-import de.maxhenkel.voicechat.VoicechatClient;
 import de.maxhenkel.voicechat.voice.client.SoundManager;
 import net.minecraft.ChatFormatting;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.network.chat.Component;
-import net.minecraft.resources.ResourceLocation;
 
 import javax.annotation.Nullable;
 import java.util.List;
 
 public class SelectSpeakerScreen extends SelectDeviceScreen {
 
-    public static final ResourceLocation SPEAKER_ICON = ResourceLocation.fromNamespaceAndPath(Voicechat.MODID, "textures/icons/speaker.png");
     public static final Component TITLE = Component.translatable("gui.voicechat.select_speaker.title");
     public static final Component NO_SPEAKER = Component.translatable("message.voicechat.no_speaker").withStyle(ChatFormatting.GRAY);
 
@@ -28,18 +23,13 @@ public class SelectSpeakerScreen extends SelectDeviceScreen {
     }
 
     @Override
-    public ResourceLocation getIcon() {
-        return SPEAKER_ICON;
-    }
-
-    @Override
     public Component getEmptyListComponent() {
         return NO_SPEAKER;
     }
 
     @Override
-    public ConfigEntry<String> getConfigEntry() {
-        return VoicechatClient.CLIENT_CONFIG.speaker;
+    public AudioDeviceList createAudioDeviceList(int width, int height, int top) {
+        return new SpeakerAudioDeviceList(width, height, top);
     }
 
 }